improve vc-got–diff and vc-got-diff
got expects arguments to diff to be relative to the worktree, it won’t
work with full path (even if those full paths are valid, i.e. inside a
worktree).

With this in place, we can also adjust vc-got-diff to process all the
files, and so now also vc-root-diff (C-x v D) works as intended.