adding vc-got-stage
vc-got-stage is a minor mode for vc-diff buffers controlled by Got.
It's a tentative at implementing the stage operation for specific
changes instead of whole files.  vc.el works with filesets, and this
is fine.  But sometimes you want to commit only certain hunks.

This initial implementation allows the user to stage changes, but you
can't commit them (yet).