add vc-got-annotate  (#4)
* add vc-got-annotate-{command,time,extract-revision-at-line}

this makes vc-annotate command basic functionality working, needs
next/prev-revision support to see diffs.

* add vc-got-{previous,next}-revision defuns

roughy working but might need a bit more testing.

* add vc-got-annotate-re helper

* use vc-got--log in prev/next-revision functions

no need to reinvent the wheel, use the existing vc-got--log to get the
log info.

* fix vc-got-previous-revision

vc-got--log leaves point at the end of buffer so move point to the
beginning before running keep-lines so it has chance at working.

* fixes for vc-got-next-revision

set rstart/end to keep-lines so it works on whole buffer and don't use
previous-line as its not intended for elisp code. Also fix the first
match regexp to allow branch names after commit sha-1 values.

* prefer with-current-buffer

* require vc-annotate

* rename vc-got-annotate-re to vc-got--annotate-re

this is for internal use so name it so. While here add doc string for it.

* Make annotation buffers parent managed by got

vc-ensure-vc-buffer won't recognise this buffer as managed by got unless
vc-parent-buffer points to a buffer managed by got.

* refactor vc-got-next-revision

no need for extensive checks in function, got command won't return
successfully unless it finds the wanted revision so are certain the rev
is found, we only need to check if it has a parent.

while here, replace the multiple commit regexp with internal
vc-got--commit-re.