fix vc-got-repository-url for checkout of non-bare repos
before I assumed got could checkout only from bare repos, but it turns
out this isn't the case.  This fixes it, making vc-got-repository-url
more robust.  It first try to use .git/config if it exists, if it
doesn't but there is a .git directory, don't do anything; finally try
to parse config (assuming this is a bare repo).