ignore ignored files
closes #12

The rationale is that `got status` doesn't print the status of the
file if it is either up-to-date or ignored by .gitignore or
.cvsignore.  Thus, we were marking as up-to-date the ignored files in
*vc-dir*.  (vc-got-state was working as expected.)

This tries to fix the situation.  File for which `got status` doesn't
print any info are saved in a double-check list, and we issue a second
`got status` only against the double check list, and push onto res
only the up-to-date ones.