annotate-time: drop save-excursion, explain why looking-at
The (save-excursion (beginning-of-line) ...) bit is not needed, as the
VC interface guarantees us that the point is at the right place before
each invocation.

While there, add an explanation for the `looking-at' usage: ideally we
should use `re-search-forward' (which moves the cursor too) but
produces an ugly result.  It's the second time I'm stumbling on this
and wasting time, so the note is warranted.