run `got fetch/send' inside the worktree
I ended up running `got fetch' and `got send' from the repo directory
when I was using `git push' to send stuff to the server.  It was the
only way.

Now, there's `got send' that works correctly and does the right thing
too when ran from a worktree, i.e. push the current branch, not
necessarily the HEAD.