fix vc-got--info command return value use
The got info command uses return value of 1 when executed for files
which are not part of got work tree which caused the emacs to show those
as errors. To avoid these errors when accessing non-registered file, tune up the
vc-got--info function to ignore the error cases when executing the
vc-got-command and only check the command return value in
vc-got-working-revision.