implement create-repo
create-repo is used to initialize a new repository for version control.
Running just "got init" does not seem sufficient as new repository must
either be served by daemon or populated with with some files. As vc is
intended for local actions the vc-got-create-repo runs "got import"
immediately on the newly created repository to populate it with initial files.