implement working-tree functionality
vc has added working-tree functionality to work with git-worktree(1).
Got does not use exact functionality but these map quite nicely to Got
work trees.

- vc-got-add-working-tree creates a new work tree with "got checkout"

- vc-got-delete-working-tree removes uuid from got repository and
  deletes the directory

- vc-got-move-working-tree just moves the directory to new place.

The git-worktree also tracks the locations of the working trees and
there is vc-known-other-working-trees command to list them. Got does not
track checkout locations so this is hard to implement currently.
When creating working-tree from vc, we could add our own file to got
repository to track these work tree locations to got some tracking to
these directories but this is left for later.