improve vc-got-checkin-patch file heuristic
the "^file" lines do not seem to give out the file when a patch is
deleting something. The '---' header seems to give the file for all
cases so use it to determine the affected file.

Add vc-got--commit-header-re to match the header block and pick the
wanted information out of it.